requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
APScheduler==3.10.4
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
Pillow==10.0.0
numpy==1.24.3 